[Test] Add business logic test spec.
